+++ examples/with-gatsby/gatsby-ssr.js
@@ -5,11 +5,3 @@
  */
 
 // You can delete this file if you're not using it
-
-/**
- * Implement Gatsby's SSR (Server Side Rendering) APIs in this file.
- *
- * See: https://www.gatsbyjs.org/docs/ssr-apis/
- */
-
-// You can delete this file if you're not using it