+++ examples/with-remix/twind.config.ts
@@ -1,7 +1,5 @@
 import { defineConfig } from '@twind/tailwind'
-import presetExt from '@twind/preset-ext'
 
 export default defineConfig({
   /* config */
-  presets: [presetExt()],
 })