+++ packages/with-remix/server.d.ts
@@ -1,0 +1,4 @@
+// ONLY FOR TYPESCRIPT TO FIND @twind/with-remix/server
+
+export { default } from './src/server'
+export * from './src/server'