+++ packages/with-remix/src/server.ts
@@ -1,0 +1,5 @@
+/* eslint-env node */
+
+export { inline as default } from 'twind'
+
+export type { InlineOptions, InlineMinify } from 'twind'